chore(components): move video-player into components sub-folder
