+++ packages/react-components/src/components/video-player/__mocks__/MockVideoPlayerProps.ts
