+++ packages/react-components/src/components/video-player/customVideoProgressBar.tsx
