+++ packages/react-components/src/components/video-player/utils/getDisplayForCustomProgressBar.spec.tsx
