+++ packages/react-components/src/components/video-player/utils/videoProgressUtils.tsx
